deployment/requirements.txt
# Flask Web Application Dependencies
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning Dependencies
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0

# Data Processing
pandas==2.0.3
scikit-learn==1.3.0

# Optional: For production deployment
gunicorn==21.2.0
python-dotenv==1.0.0

# Development dependencies (optional)
flask-cors==4.0.0
